REFACTOR: Logic for leaving Train screen was moved from exercise to Train - navigation | handleSubmit (DB) | FinishWorkout
