+++ app/screens/Train.js
@@ -1,6 +1,9 @@
 import React, { useReducer, useEffect } from 'react';
 import { Button, StyleSheet, View } from 'react-native';
 import Exercise from './Exercise';
+import Spacer from '../components/Spacer';
+import TextButton from '../components/TextButton';
+import { useDB } from '../hooks/useDB';
 // import Rest from './Rest';
 
 const MODES = {
@@ -155,6 +158,8 @@ function reducer(store, action) {
 }
 
 function Train({navigation, route:{params:{exercises, routineName, routineId}}}) {
+  const {db, workouts, timestamp } = useDB()
+
   const [store, dispatch] = useReducer(reducer, exercises, initializer);
 
   useEffect(() => {
@@ -180,11 +185,62 @@ function Train({navigation, route:{params:{exercises, routineName, routineId}}})
     )
   }
 
+  const handleSubmit = () => {
+    const millis = Date.now() - store.time;
+    const itemsCount = store.items.length;
+    const completedItemsCount = store.items.filter(item => {
+      return item.session.isFinished
+    }).length
+
+    const batch = db().batch();
+    const newRef = workouts.ref.doc(routineId).collection("routineSessions").doc();
+
+    batch.set(newRef, {
+      created: timestamp,
+      duration: millis,
+      completedItemsCount,
+    }, { merge: true });
+    
+    store.items.forEach(item => {
+      batch.set(newRef, {
+        exercises: db.FieldValue.arrayUnion({
+          exerciseName: item.exerciseName,
+          parentExercise_ref: item.parentExercise_ref,
+          childExercise_ref: item.id,
+          mode: item.mode.current,
+          [item.weight.current]: item.weight[item.weight.current],
+          isCompleted: item.session.isFinished,
+          goal: item.session.end,
+          result: item.session.count,
+        }),
+      }, { merge: true })
+    })
+
+    batch.commit();
+
+    return { duration: millis, completedItemsCount, itemsCount }
+  }
+
+  function FinishWorkout() {
+    return (
+      <View>
+        <Spacer mV={8} style={styles.line}/>
+        <TextButton onPress={() => {
+          navigation.navigate("TrainComplete", { items: store.items, routineName: routineName, stats: handleSubmit() })
+        }}>
+          finish workout
+        </TextButton>
+        <Spacer mV={8}/>
+      </View>
+    )
+  }
+
   return (
     <>
       <View style={styles.container}>
-        <Exercise navigation={navigation} store={store} dispatch={dispatch} MODES={MODES} routineName={routineName} routineId={routineId}/>
+        <Exercise store={store} dispatch={dispatch} MODES={MODES}/>
         {/* <Buttons/> */}
+        {store.index === store.items.length - 1 ? <FinishWorkout/> : null}
       </View>
     </>
   );
@@ -196,7 +252,12 @@ const styles = StyleSheet.create({
   },
   text: {
     color: 'white',
-  }
+  },
+  line: {
+    width: '100%',
+    borderTopWidth: 1,
+    borderTopColor: '#383B3B',
+  },
 })
 
 export default Train;
